+++ .gitignore
@@ -110,7 +110,4 @@ dist
 # Stores VSCode versions used for testing VSCode extensions
 .vscode-test
 
-#generated-files
-users-with-posts.json
-
 # End of https://www.toptal.com/developers/gitignore/api/node
