--- onlookers-notes-backend/nodemon.json
@@ -1,3 +1,0 @@
-{
-    "ignore": "*.json"
-}
