+++ package.json
@@ -9,7 +9,6 @@
   "author": "lucas-feitosa",
   "license": "ISC",
   "dependencies": {
-    "axios": "^0.19.2",
     "express": "^4.17.1"
   },
   "devDependencies": {
